test: LatestActivity
